Update: move files to a src directory.
