+++ src/clui/writer/consolewriter.go
