+++ src/clui/writer/consolewriter_test.go
