+++ src/clui/writer/stubwriter.go
