+++ src/game/gamefactory.go
@@ -3,9 +3,9 @@ package game
 import (
 	"regexp"
 	"strings"
-	"tictactoe/board"
-	"tictactoe/clui"
-	"tictactoe/player"
+	"tictactoe/src/board"
+	"tictactoe/src/clui"
+	"tictactoe/src/player"
 )
 
 type Factory struct {
