+++ src/player/factory_test.go
@@ -3,9 +3,9 @@ package player
 import (
 	"strings"
 	"testing"
-	"tictactoe/clui"
-	"tictactoe/clui/writer"
-	"tictactoe/testhelper"
+	"tictactoe/src/clui"
+	"tictactoe/src/clui/writer"
+	"tictactoe/src/testhelper"
 )
 
 func TestFactoryCreatesHumanPlayer(t *testing.T) {
