+++ src/player/thinkingtimer.go
